pyproject.toml
[tool.poetry]
name = "distill"
version = "0.1.0"
description = "Personal ML/AI knowledge base with epistemological rigor"
authors = ["pneirz"]
packages = [{ include = "distill" }]

[tool.poetry.dependencies]
python = "^3.12"

# LLM
anthropic = "^0.49.0"
tenacity = "^9.0.0"

# Document parsing
pymupdf = "^1.25.0"
beautifulsoup4 = "^4.13.0"
lxml = "^5.3.0"
requests = "^2.32.0"
tiktoken = "^0.9.0"

# Semantic search
faiss-cpu = "^1.9.0"
sentence-transformers = "^3.4.0"
numpy = "^2.2.0"

# Lexical search and fuzzy matching
rank-bm25 = "^0.2.2"
rapidfuzz = "^3.12.0"

# CLI and utilities
click = "^8.1.0"
pyyaml = "^6.0.2"
python-dotenv = "^1.1.0"
rich = "^13.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^6.1.0"
pytest-mock = "^3.14.0"
ipykernel = "^6.29.0"
jupyter = "^1.1.0"
ruff = "^0.11.0"

[tool.poetry.scripts]
distill = "distill.cli.main:cli"

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=distill --cov-report=term-missing"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
